[Z-002] Rename Controller file
